--- a/vllm/envs.py
+++ b/vllm/envs.py
@@ -78,7 +78,7 @@ if TYPE_CHECKING:
     VLLM_ROCM_USE_AITER_LINEAR: bool = True
     VLLM_ROCM_USE_AITER_MOE: bool = True
     VLLM_ROCM_USE_AITER_FP8_BLOCK_SCALED_MOE: bool = False
-    VLLM_ROCM_USE_AITER_FP8_CHANNEL_SCALED_MOE: bool = False
+    VLLM_ROCM_USE_AITER_FP8_TKW1_MOE: bool = False
     VLLM_ROCM_USE_AITER_RMSNORM: bool = True
     VLLM_ROCM_FP8_PADDING: bool = True
     VLLM_ROCM_MOE_PADDING: bool = True
@@ -554,11 +554,13 @@ environment_variables: dict[str, Callable[[], Any]] = {
     (os.getenv("VLLM_ROCM_USE_AITER_FP8_BLOCK_SCALED_MOE", "false").lower() in
      ("true", "1")),
 
-    # Whether to use aiter channel scaled moe kernel.
+    # Whether to use aiter tkw1 moe kernel
+    # that applies top-k weights on the first GEMM output.
+    # Currently this is used in Llama-4 only.
     # By default this is disabled.
-    "VLLM_ROCM_USE_AITER_FP8_CHANNEL_SCALED_MOE":
-    lambda: (os.getenv("VLLM_ROCM_USE_AITER_FP8_CHANNEL_SCALED_MOE", "false").
-             lower() in ("true", "1")),
+    "VLLM_ROCM_USE_AITER_FP8_TKW1_MOE":
+    lambda: (os.getenv("VLLM_ROCM_USE_AITER_FP8_TKW1_MOE", "false").lower() in
+             ("true", "1")),
 
     # use aiter rms norm op if aiter ops are enabled.
     "VLLM_ROCM_USE_AITER_RMSNORM":