--- a/vllm/model_executor/layers/fused_moe/rocm_aiter_fused_moe.py
+++ b/vllm/model_executor/layers/fused_moe/rocm_aiter_fused_moe.py
@@ -23,43 +23,43 @@ def is_rocm_aiter_channel_scaled_moe_enabled() -> bool:
         envs.VLLM_ROCM_USE_AITER_FP8_CHANNEL_SCALED_MOE
 
 
-def asm_moe_tkw1_impl(sorted_ids: torch.Tensor,
-                      sorted_weights: torch.Tensor,
-                      sorted_expert_ids: torch.Tensor,
-                      num_valid_ids: torch.Tensor,
-                      moe_buf: torch.Tensor,
-                      hidden_states: torch.Tensor,
-                      w1: torch.Tensor,
-                      w2: torch.Tensor,
-                      topk_weight: torch.Tensor,
-                      topk_ids: torch.Tensor,
-                      fc1_scale: Optional[torch.Tensor] = None,
-                      fc2_scale: Optional[torch.Tensor] = None,
-                      fc1_smooth_scale: Optional[torch.Tensor] = None,
-                      fc2_smooth_scale: Optional[torch.Tensor] = None,
-                      activation_str: str = "silu") -> None:
-    import aiter as rocm_aiter
+def rocm_aiter_asm_moe_tkw1(hidden_states,
+                            w1,
+                            w2,
+                            topk_weight,
+                            topk_ids,
+                            fc1_scale=None,
+                            fc2_scale=None,
+                            fc1_smooth_scale=None,
+                            fc2_smooth_scale=None,
+                            a16=False,
+                            per_tensor_quant_scale=None,
+                            expert_mask=None,
+                            activation_str: str = "silu") -> None:
+
+    from aiter import ActivationType
+    from aiter.fused_moe_bf16_asm import asm_moe_tkw1
 
     if activation_str == "silu":
-        activation = rocm_aiter.ActivationType.Silu
+        activation = ActivationType.Silu
     elif activation_str == "gelu":
-        activation = rocm_aiter.ActivationType.Gelu
+        activation = ActivationType.Gelu
     else:
-        activation = rocm_aiter.ActivationType.Silu
+        activation = ActivationType.Silu
 
-    E, model_dim, _ = w2.shape
-    M, topk = topk_ids.shape
-    device = topk_ids.device
-
-    a8_type = (w1.dtype if w1.dtype != torch.int32 and w1.dtype != torch.uint32
-               else torch.float8_e4m3fnuz)
-    a8 = torch.empty((M, model_dim), dtype=a8_type, device=device)
-    a8_scale = torch.empty(M, dtype=torch.float, device=device)
-    rocm_aiter.dynamic_per_token_scaled_fp8_quant(a8, hidden_states, a8_scale)
-    fmoe_func = rocm_aiter.fmoe_g1u1_tkw1
-    fmoe_func(moe_buf, a8, w1, w2, sorted_ids, sorted_weights,
-              sorted_expert_ids, num_valid_ids, topk, a8_scale, fc1_scale,
-              fc2_scale, fc2_smooth_scale, activation)
+    return asm_moe_tkw1(hidden_states,
+                        w1,
+                        w2,
+                        topk_weight,
+                        topk_ids,
+                        fc1_scale=fc1_scale,
+                        fc2_scale=fc2_scale,
+                        fc1_smooth_scale=fc1_smooth_scale,
+                        fc2_smooth_scale=fc2_smooth_scale,
+                        a16=a16,
+                        per_tensor_quant_scale=per_tensor_quant_scale,
+                        expert_mask=expert_mask,
+                        activation=activation)
 
 
 def rocm_aiter_fused_experts(
@@ -84,16 +84,10 @@ def rocm_aiter_fused_experts(
 
     from vllm.model_executor.layers.quantization.utils.fp8_utils import (
         per_token_group_quant_fp8)
-    
-    if apply_router_weight_on_input:
-        _, topk = topk_weights.shape
-        assert (
-            topk == 1
-        ), "Only support topk=1 when `apply_router_weight_on_input` is True"
 
-        hidden_states = hidden_states * topk_weights.to(hidden_states.dtype)
-        topk_ids = topk_ids.to(torch.int32)
-        topk_weights = torch.ones_like(topk_weights, dtype=torch.float32)
+    # All AITER Fused MoE kernels are expecting the following datatypes
+    topk_weights = topk_weights.to(torch.float32)
+    topk_ids = topk_ids.to(torch.int32)
 
     if is_rocm_aiter_block_scaled_moe_enabled() and use_fp8_w8a8:
         assert not apply_router_weight_on_input, (
@@ -148,53 +142,30 @@ def rocm_aiter_fused_experts(
             None,
         )
         return out_asm
-    
+
     elif is_rocm_aiter_channel_scaled_moe_enabled() and use_fp8_w8a8:
-        topk_weights = topk_weights.to(torch.float32)
-        topk_ids = topk_ids.to(torch.int32)
 
-        E, model_dim, _ = w2.shape
-        dtype = hidden_states.dtype
+        # # All AITER Fused MoE kernels are expecting the following datatypes
+        # topk_weights = topk_weights.to(torch.float32)
+        # topk_ids = topk_ids.to(torch.int32)
 
-        if expert_mask is not None:
-            E = expert_mask.numel()
-
-        (
-            sorted_token_ids,
-            sorted_weight_buf,
-            sorted_expert_ids,
-            num_valid_ids,
-            out_asm,
-        ) = rocm_aiter_asm_fmoe.moe_sorting_ck(topk_ids,
-                                               topk_weights,
-                                               E,
-                                               model_dim,
-                                               dtype,
-                                               expert_mask=expert_mask)
-        
-        asm_moe_tkw1_impl(
-            sorted_ids=sorted_token_ids,
-            sorted_weights=sorted_weight_buf,
-            sorted_expert_ids=sorted_expert_ids,
-            num_valid_ids=num_valid_ids,
-            moe_buf=out_asm,
-            hidden_states=hidden_states,
-            w1=w1,
-            w2=w2,
-            topk_weight=topk_weights,
-            topk_ids=topk_ids,
-            fc1_scale=w1_scale,
-            fc2_scale=w2_scale,
-            fc1_smooth_scale=None,
-            fc2_smooth_scale=None,
-            activation_str=activation)
-
-        return out_asm
+        return rocm_aiter_asm_moe_tkw1(hidden_states,
+                                       w1,
+                                       w2,
+                                       topk_weights,
+                                       topk_ids,
+                                       fc1_scale=w1_scale,
+                                       fc2_scale=w2_scale,
+                                       fc1_smooth_scale=None,
+                                       fc2_smooth_scale=None,
+                                       a16=False,
+                                       per_tensor_quant_scale=None,
+                                       expert_mask=None,
+                                       activation_str=activation)
 
     elif use_fp8_w8a8:
         assert not apply_router_weight_on_input, (
             "apply_router_weight_on_input is not supported for fp8_w8a8")
-        
         return rocm_aiter_asm_fmoe.asm_moe(hidden_states=hidden_states,
                                            w1=w1,
                                            w2=w2,
@@ -205,6 +176,16 @@ def rocm_aiter_fused_experts(
                                            fc1_smooth_scale=None,
                                            fc2_smooth_scale=None,
                                            a16=False)
+
+    if apply_router_weight_on_input:
+        _, topk = topk_weights.shape
+        assert (
+            topk == 1
+        ), "Only support topk=1 when `apply_router_weight_on_input` is True"
+
+        hidden_states = hidden_states * topk_weights.to(hidden_states.dtype)
+        topk_ids = topk_ids.to(torch.int32)
+        topk_weights = torch.ones_like(topk_weights, dtype=torch.float32)
 
     return rocm_aiter.ck_moe(hidden_states=hidden_states,
                              w1=w1,