… vllm/model_executor/layers/quantization/compressed_tensors/compressed_tensors_moe.py
… vllm/model_executor/layers/quantization/fp8.py